requirements.txt
numpy==1.26.2
Pillow==10.1.0
git+https://github.com/Mahdi-Alhakim/NeuralNetworkLibrary.git